[Refactor] 로또 번호 생성 기능 제작에 따른 기능 분리
LottoGame.kt, Application.kt: LottoMaker 필드 추가
processor: 로또 생성 기능 분리
