+++ src/main/kotlin/lotto/Application.kt
@@ -1,12 +1,17 @@
 package lotto
 
 import lotto.game.LottoGame
-import lotto.model.Lotto
+import lotto.model.generator.LottoMaker
+import lotto.model.generator.LottoRandomListGenerator
 import lotto.view.InputView
 import lotto.view.OutputView
 
 fun main() {
-
-    val lottoGame = LottoGame(inputView = InputView(), outputView = OutputView())
+    val lottoMaker = LottoMaker(LottoRandomListGenerator())
+    val lottoGame = LottoGame(
+        inputView = InputView(),
+        outputView = OutputView(),
+        lottoMaker = lottoMaker
+    )
     lottoGame.play()
 }
